refactor: move server to separate file
